# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Đăng Khoa Sport API
  description: Enterprise-grade e-commerce API for sports equipment.
  version: 1.0.0
  contact:
    email: support@dangkhoasport.com
servers:
  - url: http://localhost:3000/api
    description: Local Development Server
  - url: https://api.dangkhoasport.com/api
    description: Production Server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: number
        description:
          type: string
        stock:
          type: integer
        images:
          type: array
          items:
            type: string
    
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]

paths:
  /users/signup:
    post:
      summary: Register a new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, passwordConfirm]
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                passwordConfirm:
                  type: string
      responses:
        201:
          description: User created successfully
  
  /users/login:
    post:
      summary: Login user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  token:
                    type: string

  /products:
    get:
      summary: Get all products
      tags: [Products]
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: sort
          schema:
            type: string
      responses:
        200:
          description: List of products
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  results:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
    
    post:
      summary: Create a product (Admin only)
      tags: [Products]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name: 
                  type: string
                price:
                  type: number
                images:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        201:
          description: Product created

  /products/{id}:
    get:
      summary: Get product by ID
      tags: [Products]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: Product details